new compare widget (view/model based) which is updated directly by PV calback + reorganization of code for easier maintainance
