+++ src/gui/images/clock.png
