+++ src/gui/images/neq.png
