+++ src/gui/images/remove.png
