+++ src/gui/images/splitter_h.png
