+++ src/gui/images/splitter_v.png
