+++ src/gui/images/warn.png
